Extract logic in rest handlers into api module
